+++ scripts/build.mjs
@@ -20,7 +20,7 @@ const external = [
 ]
 
 const shared = {
-  entryPoints: ['src/index.ts', 'src/plugin/index.ts'],
+  entryPoints: ['src/index.ts', 'src/plugin/index.ts', 'src/spec/index.ts'],
   bundle: true,
   external,
   sourcemap: true,